msm8909: Qualcomm snapdragon bootanimation
For my own reference (how to steps)
importing bootanimation from qualcomm proprietary to devicetree(device/qcom/msm8909)
since we would like to contain the changes in our devicetree and not to change the android source code.
we also found that msm8909 is capable of running bootanimation of 720*1280 resolution at 10 frame rate.

Change-Id: I33f24d739d914aa18d967bf8d82e05a18452eb84